ARM: dts: {lisa,redwood}: Add qcom,display-panels property
Add qcom,display-panels property in order to use panel event
notifier API.

Tashar02: There's no need to add a common sde-display dtsi for
xiaomi since this is the only same change but repeated in other
dtsi file ever. We can suffice with the fragmentation. Adapt
the change separately for lisa as well.

Change-Id: I29dd39a633b95d7104c6b31b149badd041a90dfa
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>