input: pm8941-pwrkey: simulate missed key press events
The status of the keys connected to the KPDPWR_N and RESIN_N pins
is identified by reading corresponding bits in the interrupt real
time status register.  If the status has changed by the time that
the interrupt is handled then a press event will be missed.

Maintain a last known status variable to find unbalanced release
events and simulate press events for each accordingly.

Change-Id: I180469a7048a40e490f21d0f8b9504136131e75b
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>