Makefile.dtbinst: Add optional DTB_TYPES prefix
Vendor packages frequently install a set of DTBs based on a collection
of targets. For instance, DTBs which are installed as part of a target
supporting "Kalama" include the base DTBs and board DTBOs. Collecting
that list of DTBs is less ideal because the list of DTBs must be
maintained in Makefile which is responsible for compiling the DTBs plus
in another file which knows to package all those DTBs. Ideally, list can
be maintained in one Makefile.

Makefiles can specify a list of DTBs for "foo" with foo-dtbs-y. Build
script can run "make dtbs" as normal and (now) also run "make
DTB_TYPES="foo-" dtb_install" to install only the dtbs supporting foo.

Change-Id: Ic162038b03a37ba822e131a92e3a8d540194b021
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>